Convert ebay-icon to be a split component.
